consolidate some sanity stuff
this includes a sad bool switch for portable text
