+++ apps/astro-party/package.json
@@ -50,7 +50,7 @@
     "@radix-ui/react-navigation-menu": "^1.1.1",
     "@radix-ui/react-switch": "^1.0.1",
     "@sanity/block-tools": "^3.9.0",
-    "@sanity/client": "^5.4.2",
+    "@sanity/client": "^6.1.1",
     "@sanity/code-input": "^4.1.0",
     "@sanity/icons": "^2.3.1",
     "@sanity/ui": "^1.3.2",
