+++ apps/astro-party/src/utils/sanity-server.ts
@@ -1,7 +1,7 @@
 import {createClient} from '@sanity/client'
 
 export const sanityWriteClient = createClient({
-  projectId: 'z9io1e0u',
+  projectId: process.env.NEXT_PUBLIC_SANITY_PROJECT_ID,
   dataset: 'production',
   useCdn: false, // `false` if you want to ensure fresh data
   apiVersion: '2021-10-19',
