+++ apps/colt-steele/src/lib/sanity.ts
@@ -30,76 +30,6 @@ const VideoResourceSchema = z.object({
   duration: z.number().optional(),
 })
 
-type VideoResource = z.infer<typeof VideoResourceSchema>
-
-export const writeTranscriptToVideoResource = async (
-  audiofile: string,
-  order: string,
-) => {
-  const transcript = await getTranscriptText(audiofile)
-  const {srt, srtUrl} = await getSRTText(audiofile)
-
-  const getDocumentQuery = groq`*[_type == "videoResource" && castingwords.orderId == "${order}"][0]`
-
-  const document: VideoResource = VideoResourceSchema.parse(
-    await sanityWriteClient.fetch(getDocumentQuery),
-  )
-
-  const {_id, castingwords, muxAsset} = document
-
-  if (muxAsset) {
-    const {Video} = new Mux()
-    const {tracks} = await Video.Assets.get(muxAsset.muxAssetId)
-
-    const existingSubtitle = tracks?.find(
-      (track: any) => track.name === 'English',
-    )
-
-    if (!existingSubtitle) {
-      await Video.Assets.createTrack(muxAsset.muxAssetId, {
-        url: srtUrl,
-        type: 'text',
-        text_type: 'subtitles',
-        closed_captions: false,
-        language_code: 'en-US',
-        name: 'English',
-        passthrough: 'English',
-      })
-    }
-  }
-
-  return sanityWriteClient
-    .patch(_id)
-    .set({
-      castingwords: {
-        ...castingwords,
-        // We've started using plain text instead of Portable Text for the transcript
-        transcript: transcript,
-        // This is a portable text version of the transcript
-        // [
-        //   {
-        //     style: 'normal',
-        //     _type: 'block',
-        //     // TODO: Programmatically create an array of blocks for each paragraph break in the transcript
-        //     // how does this format in Portable Text?
-        //     children: [
-        //       {
-        //         _type: 'span',
-        //         marks: [],
-        //         _key: uniqueId('body-key-'),
-        //         text: transcript,
-        //       },
-        //     ],
-        //     markDefs: [],
-        //     _key: uniqueId('block-key-'),
-        //   },
-        // ],
-        srt: srt,
-      },
-    })
-    .commit()
-}
-
 const UpdateVideoResourceAssetSchema = z.object({
   sanityDocumentId: z.string(),
   castingwordsOrder: CastingWordsOrderResponseSchema,
