+++ apps/protailwind/package.json
@@ -62,7 +62,7 @@
     "@radix-ui/react-dropdown-menu": "^1.0.0",
     "@radix-ui/react-navigation-menu": "^1.1.1",
     "@radix-ui/react-switch": "^1.0.1",
-    "@sanity/client": "^3.3.2",
+    "@sanity/client": "^6.1.1",
     "@sanity/webhook": "^1.0.2",
     "@sentry/nextjs": "^7.9.0",
     "@sentry/tracing": "^7.9.0",
