+++ apps/testing-javascript/package.json
@@ -32,7 +32,7 @@
     "@radix-ui/react-navigation-menu": "^1.1.1",
     "@radix-ui/react-switch": "^1.0.1",
     "@sanity/block-tools": "^3.11.1",
-    "@sanity/client": "^4.0.1",
+    "@sanity/client": "^6.1.1",
     "@sanity/code-input": "^3.0.1",
     "@sanity/icons": "^2.2.2",
     "@sanity/ui": "^1.0.12",
