+++ apps/testingaccessibility/package.json
@@ -68,7 +68,7 @@
     "@reach/dialog": "^0.16.2",
     "@reach/skip-nav": "^0.16.0",
     "@sanity/block-content-to-react": "^3.0.0",
-    "@sanity/client": "^3.2.0",
+    "@sanity/client": "^6.1.1",
     "@sanity/webhook": "^1.0.2",
     "@sentry/nextjs": "^7.9.0",
     "@sentry/tracing": "^7.9.0",
