+++ apps/transcript-automation/package.json
@@ -22,7 +22,7 @@
     "@next-auth/prisma-adapter": "^1.0.3",
     "@next/mdx": "^13.4.3",
     "@portabletext/react": "^1.0.6",
-    "@sanity/client": "^3.2.0",
+    "@sanity/client": "^6.1.1",
     "@sanity/webhook": "^1.0.2",
     "@sentry/nextjs": "^7.9.0",
     "@sentry/tracing": "^7.9.0",
