+++ apps/typescriptcourse/package.json
@@ -45,7 +45,7 @@
     "@reach/dialog": "^0.16.2",
     "@react-three/drei": "^7.20.2",
     "@react-three/fiber": "^7.0.27",
-    "@sanity/client": "^3.3.2",
+    "@sanity/client": "^6.1.1",
     "@sanity/webhook": "^1.0.2",
     "@sentry/nextjs": "^7.9.0",
     "@skillrecordings/analytics": "workspace:*",
