+++ packages/create-skill-app/templates/next/src/utils/sanity-server.ts
@@ -1,7 +1,7 @@
 import sanityClient from '@sanity/client'
 
 export const sanityWriteClient = sanityClient({
-  projectId: 'z9io1e0u',
+  projectId: 'NO',
   dataset: 'production',
   useCdn: false, // `false` if you want to ensure fresh data
   apiVersion: '2021-10-19',
