+++ packages/skill-api/package.json
@@ -33,7 +33,7 @@
   },
   "dependencies": {
     "@babel/runtime": "^7.16.3",
-    "@sanity/client": "^4.0.1",
+    "@sanity/client": "^6.1.1",
     "@skillrecordings/commerce-server": "workspace:*",
     "@skillrecordings/convertkit-react-ui": "workspace:*",
     "@skillrecordings/database": "workspace:*",
