+++ packages/skill-lesson/package.json
@@ -23,6 +23,7 @@
     "@code-hike/mdx": "^0.8.2",
     "@heroicons/react": "^1.0.6",
     "@mdx-js/mdx": "^2.3.0",
+    "@mux/mux-node": "^6.3.0",
     "@mux/mux-player-react": "1.8.0",
     "@portabletext/react": "^1.0.6",
     "@portabletext/types": "^1.0.3",
@@ -32,7 +33,7 @@
     "@radix-ui/react-label": "^2.0.2",
     "@radix-ui/react-slot": "^1.0.2",
     "@radix-ui/react-switch": "^1.0.1",
-    "@sanity/client": "^4.0.1",
+    "@sanity/client": "^6.1.1",
     "@skillrecordings/ability": "workspace:*",
     "@skillrecordings/analytics": "workspace:*",
     "@skillrecordings/commerce-server": "workspace:*",
